requirements.txt
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.6
django-bagou==0.1.0
django-cors-headers==3.1.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Flask==1.0.2
gunicorn==20.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
phoenixdb==0.7
pika==0.9.13
Pillow==6.2.1
protobuf==3.10.0
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tool-yy==0.1.4
tornado==3.1.1
urllib3==1.24.3
Werkzeug==0.15.5
whitenoise==4.1.4
